feat: splits working code
