+++ apps/web/src/components/editor/timeline/index.ts
@@ -1,0 +1,11 @@
+export { Timeline } from "../timeline";
+export { TimelineToolbar } from "./timeline-toolbar";
+export { TimelineContent } from "./timeline-content";
+export { TimelineRuler } from "./timeline-ruler";
+export { TimelineTracksArea } from "./timeline-tracks-area";
+export { TrackIcon } from "./track-icon";
+export { useTimelineDragHandlers } from "./timeline-drag-handlers";
+export { useTimelineActionHandlers } from "./timeline-action-handlers";
+export { useTimelineScrollSync } from "./timeline-scroll-sync";
+export { useTimelineContentClick } from "./timeline-content-click";
+export { useTimelineWheelHandler } from "./timeline-wheel-handler";
